change db to mongoDb and add infrastructure for dynamic including based on strategy.

+++ src/Services/Posts/InstaConnect.Posts.Domain/Features/Posts/Models/Responses/PostCollection.cs
@@ -5,6 +5,6 @@ public record PostCollection(
     ICollection<Post> Data,
     int Page,
     int PageSize,
-    long TotalCount,
+    int TotalCount,
     bool HasNextPage,
     bool HasPreviousPage);
